Do not filter out issues matching exclude path
Previously, if a check reported an issue in a package that matched
the godel "exclude" directive, it was ignored. However, if a check
reports an issue in an excluded package despite the package not
being provided to the check, there is a good likelihood that the
issue is pertinent and should not be ignored.

Fixes #103