Revert "Do not filter out issues matching exclude path (#104)"
Restore previous behavior of skipping issues that match exclude
criteria. The reasoning for changing this behavior was incorrect:
the thought was that the exclude list controlling the input packages
would be sufficient. However, exclude criteria can also match at the
level of individual files, and there is no good way to support this
behavior. Because the original PR being reverted by this commit was
made to enable specific edge-case behavior (showing failures for
"compiles" check even if errors are reported in paths that are
specified as ignore paths), it is not worth breaking this general
case.

This reverts commit e7caa5d881615fb7b828a404a7178ae34b36b729.